[DEV-9734b] Handle excluded data
